fix: #1189, checkbox e.nativeEvent
